update all of `linkerd2-proxy-http` to `std::future` (#541)
This PR updates everything remaining in the `linkerd2-proxy-http` crate
to use `std::future`. In particular, this includes the following
modules:

* `add_header`
* `insert`
* `override_authority`
* `header_from_target`

The updates for these modules was pretty mechanical and there are no
significant changes, just updates to use the new versions of the
`Future` and `Service` traits.

In addition, I've put back the stack layers from these modules, where
possible (in some cases, these modules were used by other code that has
yet to be updated or is pending in a separate branch). I've also
re-enabled some of the the `discovery` integration tests for the
`linkerd2-dst-override` header. Some of these tests depend on service
profiles as well, so they will be enabled once #537 merges.

Furthermore, I've done a little cleanup now that everything in the crate
has been updated. I removed the `grpc` module entirely. Its only purpose
was converting between the `tower_grpc::Body` and `hyper::Payload`
traits, and `tonic` and the `std::future` version of `hyper` both use
the `http_body::Body` trait, so this conversion is unnecessary. Finally,
since all the code which depended on the old versions of `tokio` and
`futures` has now been updated, I've removed the dependencies on the
legacy versions, and renamed the `futures_03` and `tokio_02` renaming
imports to just `futures` and `tokio`.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>