propagate spans to tasks spawned by Hyper
Previously, in the `futures` 0.1 version of the proxy, we provided
`hyper` with a custom executor to propagate `tracing` spans to tasks
spawned by `hyper`. In the process of updating the proxy to
`std::future`, this was removed, meaning that any tasks spawned by hyper
are missing their spans. This makes events recorded by `h2` and so on
hard to debug, especially when many connections are open.

This branch fixes this by adding an implementation of `hyper::Executor`
that propagates the current span when spawning a task.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>