Add a CODEOWNERS
This CODEOWNERS change will enforce a requirement that I approve all
PRs to master. This is already practically the case, but it's not
codified.

I'd like to follow this by relaxing the PR review requirement to 1. This
allows me to merge my own PRs with only a single review, and will allow
me to unblock any other PR. Practically, I plan to continue to solicit
>1 review on changes; but I don't think this needs to be enforced by
GitHub.