make shutdowns less graceless (#560)
Although the `Drain` type which powers the proxy's graceful shutdown
machinery was updated in #482, we were not actually triggering graceful
shutdowns. This is because the Hyper `ConnectionFuture` was wrapped in a
compatibility adapter which did not allow access to the inner future so
that we could call the `graceful_shutdown` method on it. When we updated
to the `std::future` version of Hyper, the graceful shutdown code was
not re-enabled.

This branch puts back the call to `ConnectionFuture::graceful_shutdown`.
It also re-enables the shutdown integration tests, which pass now.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>