move DNS resolver back to the admin thread (#559)
Upstream changes to `trust-dns-resolver` removed the background task to
drive DNS resolutions. Now, the resolution is performed as part of the
lookup future.

This means that resolutions are now occurring "inline" when we call into
a resolver, rather than on the background thread. This caused issues
with failed resolutions preventing the proxy from shutting down.

This should hopefully fix that.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>