re-enable proxy integration tests (#561)
This branch re-enables all the proxy integration tests that hadn't
already been turned back on. 

It turns out that almost everything just works!

Depends on #560

Signed-off-by: Eliza Weisman <eliza@buoyant.io>

* remove ENOTCONNs from HTTP transport metrics tests

The way Hyepr handles disconnections has changed, so we no logner see
this errno. This is actually more consistent with the way clean
disconnections are handled for raw TCP traffic, anyway!

Signed-off-by: Eliza Weisman <eliza@buoyant.io>

* put back upgrade service that i apparently forgot

Signed-off-by: Eliza Weisman <eliza@buoyant.io>

* cargo.lock

* fix zero length `read`s in transparency tests

whoops --- `Vec::clear` sets the length to 0, so no bytes are read into
the buffer. fixed this by just allocating a new vec like the tests used
to do, it's fine, they're just tests.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>

Co-authored-by: Oliver Gould <ver@buoyant.io>