Use async/await to simplify connection-accept task (#575)
The `app::core::serve` provides a task that drives connections from a
listener to an accept stack until a drain is signaled, at which point
the task completes and no further connections are accepted.

The implementation for this can be simplified by using newer async/await
syntax. The `drain::Watch` API has been improved to support this.