outbound: Split HTTP endpoint builder
This change follows on the refactor in d7b86b01 by splitting the
outbound HTTP endpoint stack into its own constructor in order to
modularize the proxy stacks. The response body is boxed in order to hide
the stack's details (with regard to metrics, tracing, etc).

There is no functional change.