Eliminate Bind & Listen traits
We have some abstractions, like the `Bind` and `Listen` traits, that
don't really serve a purpose. Each have only a single implementation in
our codebase. And, furthermor, the `Listen` trait is better-represented
as a `Stream` (so that we can use general tools with it).

This change eliminates these traits and modifies the implementation to
use `async-stream` to take advantage of `async/await` syntax. Similarly,
the logic of `listen::Serve` is now directly implemented in the `serve`
task. The `proxy::cor::listen` module has been replaced by an `accept`
module.

All of this helps us cut down on unnecessary types and helps to
consolidate logic (rather than having it spread across many crates).