update test-support clients and servers to be natively async (#580)
Currently, the test clients and servers in linkerd2-app-integration
each run their own separate Tokio runtime on a background thread. This
was done in order to allow writing the test code in a synchronous
fashion, rather than using futures. Now, however, a majority of the
tests already use async-await syntax, so we can run all this stuff on
the same test runtime. This should reduce the number of threads in the
tests significantly, which may decrease flakiness and possibly improve
test performance a bit.

This branch updates the test support servers and clients to run as tasks
spawned on the main test runtime. It does not touch the test control
plane (Destination and Identity services) as they still need to be
updated to use Tonic and std::future.

This does introduce one potential footgun that is worth noting: if a
test server that makes assertions is spawned in the background, the
JoinHandle for its task must be awaited. Otherwise, panics will not be
propagated to the main test task, and the test can spuriously pass if
the assertion fails.

Signed-off-by: Eliza Weisman eliza@buoyant.io