Simplify tap server (#582)
## Description

Remove the Tap daemon that is responsible for handling service registrations and
tap subscriptions.

There is now a registry that contains all active taps.

The tap service has access to this registry and it uses it to retrieve an
updated set of active taps for every request. This set is used to determine
whether the request should be tapped.

New taps are added when the gRPC server receives an observe request. If the
request is valid, the server adds the tap to the registry.

Inactive taps are cleared by a background task. This background task is much
simpler than the one before; it loops on an interval and checks whether each tap
still has an actively used response stream, and if it has not reached its limit.

### Changes

The primary change is the addition of the `Registry` in
`linkerd/proxy/tap/src/registry.rs`.

The tap service holds on to a `Registry` and uses the `taps_recv` field to get
an updated list of taps for each request. The tap service is not responsible for
adding or removing any taps from the registry.

The gRPC service also holds on to a `Registry` and uses the `inner` field to
register new taps with the registry. The field is locked giving access to an
`Inner`. With this, the gRPC server can add new taps to the registry and then
send updates with the `taps_send` sender.

Neither the tap service or the gRPC is responsible for clearing out inactive
taps. To solve this, we spawn a background task with the `async fn clean` method
which loops on an interval and checks that each tap still has an actively used
response stream.


### Traits

The `tap::iface` module now only contains the `Tap`, `TapPayload`, and
`TapResponse` traits. These are helpful for traits that are implemented in
`tap::grpc`.

The `Subscribe` and `Register` traits have been removed. Services no longer
*register* with the tap server. Taps still do subscribe (now called `register`)
with the registry, but it was only implemented by one object and did not
simplify the implementation in terms of understanding or save on duplicated
code.

Signed-off-by: Kevin Leimkuhler <kevin@kleimkuhler.com>