Add `multicore` to default features
We don't have any blockers to using the threaded runtime by default.
This change adds the `multicore` feature flag to the default feature
list.

It also modifies the threaded runtime initialization to use all
available cores.

The thread names have been changed from `linkerd2-proxy-worker` to just
`proxy`, which is more consistent with the `admin` thread name. This
should be unambiguous in the context of the process.