trace: tweak tracing & test support for the multithreaded runtime (#616)
This branch makes a couple minor tweaks to improve tracing with the
multithreaded Tokio runtime. 

It enables the option to print thread IDs when logging trace output,
which should make it somewhat clearer when events occur on different
threads.

Additionally, there's currently an issue where traces are not logged
from the proxy during integration tests. Only traces from the test
support code are logged. This is because we are using `tracing`'s scoped
thread-local dispatcher in the test proxy, so that each test can create
its own subscriber. We set the test's subscriber as the default for the
test proxy thread, but *not* as the global default for all threads.
However, when the multithreaded runtime is enabled, the default runtime
returned by `tokio::runtime::Runtime::new` is now a multithreaded
runtime. Therefore, the test proxy runs on the runtime's worker threads,
rather than the main thread spawned by the tests, and the dispatcher is
not set for the worker threads. 

This branch fixes that issue by explicitly using the basic (single
threaded) scheduler when creating the test proxy. This should
(hopefully) also improve test flakiness a bit by not spawning a whole
bunch of workers.