Introduce a SkipDetect layer to preempt detection (#620)
This change introduces a new SkipDetect layer that configures whether
protocol detection should be attempted. This module will later be
replaced/augmented by discovery.

Furthermore, this change eliminates the `Accept` trait. Instead of
modeling the accept stack as a simple service whose response is a
future, we know model the stack as, effectively, a
`MakeService<Meta, TcpStream>`. This is intended to support caching of
the service that handles the tcp stream (i.e. to hold discovery
responses).

Detection timeouts have been moved from a dedicated layer into the
detection modules.