Remove `Resolution` trait (#606)
The `Resolution` trait was initially introduce to ensure that the update stream
is infinite; but this isn't actually desirable in a general way. It's fine for resolution
streams to complete. Consumers (balancers) need to handle this.

This change removes this trait, in favor of a simple `TryStream`. This makes it
easier to use ready-made stream combinators and constructors.