Split orig-proto tests out of discovery tests (#629)
About 18 months ago, we removed support for a variety of headers that
could potentially leak information to external clients and services; but
the implementation and tests remain in place.

In the interest of removing cruft, this change removes these unused
modules (which are not compiled, and may therefore not even compile
today). These headers are: `l5d-client-id`, `l5d-server-id`, and
`l5d-remote-ip`.