resolve: Eagerly fail resolutions
In 591191c6, the initial resolution was moved into the stream. This
prevents the initial future from failing with an error when the control
plane fails eagerly (as is the case with InvalidArgument errors).

To preserve the eager failure, the stream is not returned until an
initial resolution has been completed.