dns: Run DNS resolutions on the main runtime
DNS resolutions are run on the admin runtime. This requires an
unnecessary layer of indirection around the resolver, including an MPSC.

Now that we allow the main runtime to use more than one thread, it's
preferable to do this discovery on the main runtime and we can simplify
the implementation.