Unify control plane client construction
We construct three control plane clients. All of the clients are
basically the same (at least with regard to the stack/configuration).
Rather than duplicate this logic throughout the code, we can unify it in
the `control` module (hiding the details of how the client is built).

This change modifies how the OpenCensus collector client is
instantiated. The `SpanExporter` no longer constructs a new client each
time it wishes to initiate a stream, opting to just clone a single
client instead.

This change also fixes reconnect logic for these clients. When the
control client balancer was introduced, the reconnect layer incorrectly
wrapped the balancer (instead of the endpoint). This has been corrected.