outbound: Cache balancers within profile stack (#641)
In an upcoming change, we'd like to do service profile discovery within
the TCP-accept stack (and not the HTTP routing stack). But, today,
service profile discovery is tightly coupled to the HTTP middleware
implementations. This change splits the service profile layers (profile
discovery, http request profiles, and traffic splitting) into several
layers, so that the discovery logic is decoupled from the http-specific
request-routing middleware.

This change removes the balancer cache and the balancer-specific buffer
so that balancers are owned by the split layer. The buffer has been
moved outside of the split layer to drive all balancers in a split and
to make the split cloneable (for the retry middleware). All of this is
cached under the profile cache.

Breaking changes:

This likely breaks support for "external" service profiles, where the
proxy resolves service profilse for domains that it cannot resolve via
the resolver. This feature is not enabled by default and is considered
experimental. We'll have to take care to at least document this in
upcoming releases.

Side-effects:

* Latency improvements at concurrency ![image](https://user-images.githubusercontent.com/240738/91736211-2fd48280-eb62-11ea-8420-30e0242980a5.png)
* Compile time improved by 20-25% (integration tests in CI ~20m => ~15m);