profiles: Do not rely on tuples as stack targets
Using tuples for stack targets (particularly, the input targets of the
route_request and split modules) is a bit brittle.  Instead, we can use
`AsRef` on the target type to access the profile receiver.

This change introduces new target types to be used to satisfy these
traits, and generally cleans up stack construction.