http: Simplify stacks and target types
Currently, various properties of HTTP requests are exposed into the
stack target type, meaning that stacks (including their resolutions)
cannot be shared across requests that vary by any of these settings and,
worse, that per-request determinations must be made to determine which
stack to use.

In order to perform bind stacks according to connection-level metadata,
where these request-level parameters cannot be known, we need to remove
this metadata from the stack target.

The HTTP server now applies URI normalization as an internal
implementation detail (because it's a detail of how hyper clients and
servers differ). This eliminates the `ShouldNormalizeUri` trait, as that
determination can be made entirely based on the properties of the
request without any additional stack-specific metadata.

The HTTP client now takes an updated `Settings` type that explicitly
handles the orig-proto upgrade logic (removing this from the client
stack).

Stack targets now hold an `http::Version` on the accept-side, and map
that to a per-endpoint `http::client::Settings` for use by the http
client.