Make SkipDetect more generic as stack::MakeSwitch
SkipDetect is needlessly coupled to its `listen::Addrs` target type and
the Accept stack's shape. In reality, it's pretty similar to the
fallback module, except that decisions are made on the target rather
than the response future.

This change replaces the `core::proxy::skip_detct` module with
`stack::switch`.

Above all, this allows us to use other targets with this module, which
is needed to do discovery before protoocl detection (and to use the
results of discovery to inform detection).