tracing: Elide redundant info in tracing contexts
The tracing context includes redundnant information, particularly about
the traffic target.

This change modifies the accept stack to include both the source peer and
target addresses in the `accept` context; and the target address has
been removed from intermediate contexts, many of which can be moved to
the debug level now.