stack: Make NewService mutable
The `NewService` trait takes an immutable reference when calling, but
this makes it insuitable for layers like `Cache`, but there is no
fundamental limitation that requires limiting the trait to be immutable.

This change modifies `NewService` to take a mutable reference so that
the trait may be used in more contexts.