cache: Implement NewService (#668)
The cache used to implement logic in its `poll_ready`; but this was
changed some time ago. Otherwise, the cache is synchronous and
infallible, much like the `NewService` trait.

This change modifies `NewService` to take a mutable reference, as needed
by the cache; and it modifies the cache to implement `NewService`
instead of `Service`.

The stacks then coerce the cache back into a service (via `into_make`)
to satisfy the buffer's requirements. It is expected that this buffering
will be decoupled from the cache in a followup change.