Use a single HTTP cache in each direction
Currently, all of the fallback logic is implemented outside of the
cache, and then each stack must be cached individualy.

This change consolidates all caching in each direction to a single
cache instance.

This is accomplished by coercing more of the stack to be implemented as
a `NewService`, rather than `MakeService`.

`ResultService` is introduced so that `Admit` can operate on
`NewService`.