Split prefix buffering from HTTP detection (#676)
This change modifies `serve` to take a `NewService` instead of a
`MakeService`. Services specific to the accept stack have been
updated as well

`DetectHttp` has been updated to work as either a `MakeService` or
`NewService` -- the asynchronous version is still needed by the outbound
proxy (until caching is changed).

`DetectTls` is now purely a `NewService`.