outbound: Remove required identity from HttpLogical (#683)
Presently, the proxy extracts the required identity from the
l5d-required-identity header into the HttpLogical target type when
it is present, and builds endpoint stacks which require that identity.
This was previously necessary because service discovery couldn't be
performed for pod IPs, only for authorities. However, this is no longer
the case, and we can now just use the identity returned by service
discovery for these stacks. This means the required identity can be
removed from the HttpLogical target, and identities in HttpEndpoint
targets are only provided by the destination metadata.

Fixes linkerd/linkerd2#5029