stack: add SwitchReady service (#694)
This is essentially a form of fallback that falls back when a primary
service takes too long to become ready. When the primary service's
`poll_ready` does not return ready for over the specified duration, the
secondary service is used until the primary becomes ready again,
resetting the readiness timeout. Erros from both services are
propagated.

I considered naming this `ReadyFallback`, but that name was already used
for an unrelated type in the `fallback` module. :)

Signed-off-by: Eliza Weisman <eliza@buoyant.io>