fix cloning detecthttp blowing away state (#700)
The `AcceptHttp` service is stateful. It builds the underlying HTTP,
HTTP/2, and TCP service stacks lazily the first time they're needed.
This interacts badly with the derived implementation of `Clone`: if it's
cloned before the state is set, the state will only be set for the
clone, rather than for the original instance in the cache. Since
`Prefix` clones and oneshots the underlying service (`AcceptHttp`), this
means that we rebuild the HTTP, HTTP/2, and TCP stacks on every
connection to the same orig dest. This basically breaks caching for
everything other than service profile discovery, causing us to do new
destination discovery for every connection to the same original
destination. Which...isn't great.

This branch fixes this issue by removing the `Clone` impl from
`AcceptHttp` and putting it behind a buffer instead. In the future, it
may be possible to fix this in a nicer way, but this resolves the issue
for now.