outbound: Unify TCP & HTTP target types
We have target types like `{Http,Tcp}{Logical,Concrete,Endpoint}` that
differ only superficially between the HTTP and TCP variants. This
results in unnecessary, duplicate trait impls.

This change consolidates these targets as `Logical`, `Concrete`, and
`Endpoint`, all of which are generic over the protocol type. We provide
`Tcp*` and `Http*` trait aliases as a shorthand and compatibility.