inbound: Fix URI normalization for orig-proto requests
linkerd/linkerd2#5069 repors that the proxied requests can incorrectly
end up in the absolute-form. This happens because URI normalization is
applied after protocol downgrade (where the absolute-form is annotated
explicitly).

This change improves the inbound orig-proto test to detect this bug and
fixes it by performing normalization before protocol downgrade so that
the downgrade layer's behavior is not overridden.