outbound: Set source address in Tap
We recently stopped exposing source addresses on outbound Tap metadata,
but this causes some confusion in the Tap UIs.

This change introduces an `io::PeerAddr` trait that is implemented by
all `Io` types and a `proxy::http:SetClientAddr` middleware that inserts
a `ClientAddr` type onto requests. This middleware is now applied by the
`DetectHttp` server to ensure the HTTP stack can access the source
address.