inbound: use MakeSwitch for loopback
Currently, loopback connections are detected using the `PreventLoop`
layer, which returns an error when the target address' port matches the
proxy listener's port. The inbound router currently uses a `Fallback`
layer to detect this error and send loopback traffic to a loopback
service (the mutlicluster gateway).

This branch changes the inbound proxy to dispatch loopback connections
to the multicluster gateway using a `MakeSwitch` layer instead. The
`MakeSwitch` layer inspects the target type and applies a predicate to
it that determines which of two services to make. This is essentially
the same behavior as using the fallback on error layer, but with fewer
steps.