Reduce tracing spans to the debug level
We currently set almost all span metadata at the info level, meaning
that these spans are recorded/maintained for all connections/requests
even though there are virtually no logs emitted with this metadata at
the info level.

This changes all of these spans to only be recorded at the debug level.
The one exception is that the per-connection `accept` span is maintained
at the info level so that errors are reported with at least
connection-level metadata.