buffer: Remove readiness watch
Previously, our `buffer` implementation exposed the inner service's
(initial) readiness via a `watch`. This was primarily present due to the
`fallback` layer which used the service's initial readiness to determine
whether to use the service -- on failures it would fallback to another
service.

Now that fallback is no longer used, the overhead of this readiness
watch can be removed.