http: Configure client connection pools
When an HTTP/1.1 client has many concurent requests, it may create an
arbitrary number of TCP connections. These connections are retained for
90s before they are dropped. In HTTP/1.1. applications communicate at
high concurrency, this can lead to elevated memory requirements.

Similar behavior is observed with HTTP/2 clients (i.e. when using mesh
protocol upgrades) if the application communicates with many unique
services, as is the case for Prometheus.

To address this issue, we want to be more aggressive about tearing down
idle/unused clients. This is primarily achieved by reducing the outbound
"max idle age" setting to 5s, which matches the default Kubernetes DNS
TTL. On the inbound side, the max idle age is set to 20s, as the number
of clients is typically on the order of <5.

Additionally, `MAX_IDLE_CONNS_PER_ENDPOINT` configurations are
introduced to bound the number of idle connections that may be retained
in an HTTP/1.1 client's connection pool. These pools remain unbounded by
default, as this setting causes elevated CPU usage in load tests. It is
generally preferable to bound the length of time an idle
client/connection is cached.