outbound: Skip endpoint resolution on profile hint
Profile resolutions may provide an endpoint response when the
target is a pod or otherwise known endpoint. When these endpoints fail
requests with an I/O error, we assume that the endpoint is defunct and
fail the service so that subsequent connections require a new
resolution.

Addresses https://github.com/linkerd/linkerd2/issues/5209