io: Set a minimum detect buffer size
If a client writes a partial message, `io::Peek` may return incomplete
data. This change modifies the `io::Peek` future to have a minimum
buffer size.

For HTTP detection, we require enough data to check the HTTP/2 magic
header (which is also roughly the same size as the smallest possible
HTTP/1.1 request.

This is primarily necessary for upcoming server-first TCP mTLS wrapping.