reconnect: Allow recovery to vary by target
The reconnect layer accepts a uniform recovery strategy for all targets.
But, now we have a need to only reconnect on certain endpoint stacks.

This change modifies the reconnect layer to take a `NewService` that
produces a `Recover` based on the target. It also modifies its connect
stack to work similarly, by binding the inner service to a target via a
`NewService` that produces a `Service<()>`, which is called each time a
connection is required.