outbound: Fail HTTP clients when any request fails
HTTP request failures (i.e. requests that do not actually receive an
HTTP response) typically indicate some deeper underlying issue with the
client. It's _also_ possible, for HTTP/1.1 requests, that a single
request encountered some socket-level issue, i.e. due to connection
pooling interacting with application-level idle timeouts.

This change modifies the outbound endpoint stack to assume that these
failures are generally terminal and that a new client should be procured
if appropriate. In-flight requests should not be impacted by this
change.