cache: Make the cache cloneable with RwLock (#743)
Caches are typically wrapped with buffers so that they can be shared
across connections. These buffers can be eliminated if the cache itself
becomes cloneable; and this is easily achieved with a
`std::sync::RwLock`: the cache first obtains a read lock to clone the
inner service and a write lock is only obtained when the service is not
in the cache.

It turns out that the buffers have an important side effect, though: the
buffers hide the inner service type, which drastically reduces the
amount of memory needed to compile the linkerd2-proxy crate. To
replicate this, this change introduces a `stack::BoxNewService` which
can be used to hide the type signature of a `NewService`.

Benchmarks indicate that this improves p50 latency by 2ms for HTTP/2
traffic at 100 concurrent requests.