http: Teardown serverside connections on error
When the proxy encounters an error, it's usually the case that we want
to gracefully teardown the serverside connection so that the
application's client has an opportunity to re-resolve the endpoint
before reconnecting.

This change configures HTTP servers so that the error-handling layer
initiates server shutdown when an error is not a request timeout. Socket
errors, failfast errors, etc, are now met with a 502 Bad Gateway, as
they were before; but after in-flight requests are responded to, the
proxy's server closes its connection with the application client.