concurrency-limit: Drop permit on readiness
The concurrency-limit module holds permits until the response future is
dropped. This is fine in theory, but we have reports of clients hitting
the concurrency limit unexpectedly.

This change ensures that the permit is released immediately as soon as
the inner future becomes ready, regardless of how long the response
future is held. This also adds trace logging when the permit is
released.