Increase the default buffer capacity to 100K (#752)
We've received reports that the default in-flight limit may be too
small. Now that the proxy can consume multiple cores, it seems wise to
increase the default from 10K to 100K.

This bound should prevent high-traffic workloads from hitting the
in-flight limit in all but extreme cases, while still protecting the
proxy against infinite growth.